+++ Documentation/python/tdw_utils.md
@@ -583,70 +583,6 @@ _Returns:_  A numpy array: `[r, p, y]` angles in radians.
 
 ***
 
-## `AudioUtils`
-
-`from tdw.tdw_utils import AudioUtils`
-
-Utility class for recording audio in TDW using [fmedia](https://stsaz.github.io/fmedia/).
-
-Usage:
-
-```python
-from tdw.tdw_utils import AudioUtils
-from tdw.controller import Controller
-
-c = Controller()
-
-initialize_trial()  # Your code here.
-
-# Begin recording audio. Automatically stop recording at 10 seconds.
-AudioUtils.start(output_path="path/to/file.wav", until=(0, 10))
-
-do_trial()  # Your code here.
-
-# Stop recording.
-AudioUtils.stop()
-```
-
-***
-
-#### `get_system_audio_device() -> str`
-
-_This is a static function._
-
-_Returns:_  The audio device that can be used to capture system audio.
-
-***
-
-#### `start(output_path: Union[str, Path], until: Optional[Tuple[int, int]] = None) -> None`
-
-_This is a static function._
-
-Start recording audio.
-
-| Parameter | Description |
-| --- | --- |
-| output_path | The path to the output file. |
-| until | If not None, fmedia will record until `minutes:seconds`. The value must be a tuple of 2 integers. If None, fmedia will record until you send `AudioUtils.stop()`. |
-
-***
-
-#### `stop() -> None`
-
-_This is a static function._
-
-Stop recording audio (if any fmedia process is running).
-
-***
-
-#### `is_recording() -> bool`
-
-_This is a static function._
-
-_Returns:_  True if the fmedia recording process still exists.
-
-***
-
 ## `QuaternionUtils`
 
 `from tdw.tdw_utils import QuaternionUtils`
