+++ Python/example_controllers/audio/rube_goldberg_objects.json
@@ -1,0 +1,315 @@
+{
+  "0": {
+    "library": "models_special.json",
+    "model_name": "prim_sphere",
+    "position": {
+      "x": -12.95,
+      "y": 1.725,
+      "z": -4.2
+    },
+    "rotation": {
+      "x": 0,
+      "y": 90,
+      "z": 0
+    },
+    "scale": {
+      "x": 0.1,
+      "y": 0.1,
+      "z": 0.1
+    }
+  },
+  "1": {
+    "library": "models_core.json",
+    "model_name": "bench",
+    "position": {
+      "x": -12.95,
+      "y": 0.96,
+      "z": -5.93
+    },
+    "rotation": {
+      "x": 0,
+      "y": 90,
+      "z": 0
+    },
+    "scale": {
+      "x": 1.25,
+      "y": 1.0,
+      "z": 1.15
+    },
+    "physics": {
+      "kinematic": true,
+      "dynamic_friction": 0.4,
+      "static_friction": 0.4,
+      "bounciness": 0.6,
+      "mass": 200
+    }
+  },
+  "10": {
+    "library": "models_core.json",
+    "model_name": "square_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3719,
+      "z": -6.273
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 0.5
+    }
+  },
+  "11": {
+    "library": "models_core.json",
+    "model_name": "square_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3927,
+      "z": -6.1707
+    },
+    "rotation": {
+      "x": -12.2,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 0.5
+    }
+  },
+  "12": {
+    "library": "models_core.json",
+    "model_name": "star_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3719,
+      "z": -6.338
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 0.5
+    }
+  },
+  "13": {
+    "library": "models_core.json",
+    "model_name": "rectangle_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3719,
+      "z": -6.411
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 0.5
+    }
+  },
+  "14": {
+    "library": "models_core.json",
+    "model_name": "l-shape_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3719,
+      "z": -6.481
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 0.5
+    }
+  },
+  "2": {
+    "library": "models_core.json",
+    "model_name": "wood_board",
+    "position": {
+      "x": -12.95,
+      "y": 1.481,
+      "z": -5.696
+    },
+    "rotation": {
+      "x": -12.93,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 2,
+      "z": 1
+    },
+    "physics": {
+      "kinematic": false,
+      "dynamic_friction": 0.2,
+      "static_friction": 0.2,
+      "bounciness": 0.7,
+      "mass": 0.6
+    }
+  },
+  "3": {
+    "library": "models_core.json",
+    "model_name": "camera_box",
+    "position": {
+      "x": -12.95,
+      "y": 1.483,
+      "z": -5.0
+    },
+    "rotation": {
+      "x": 0,
+      "y": -90,
+      "z": 90
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 1
+    },
+    "physics": {
+      "kinematic": true,
+      "dynamic_friction": 0.45,
+      "static_friction": 0.48,
+      "bounciness": 0.6,
+      "mass": 0.5
+    }
+  },
+  "4": {
+    "library": "models_core.json",
+    "model_name": "toy_monkey_medium",
+    "position": {
+      "x": -12.95,
+      "y": 1.591,
+      "z": -5.1
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 1
+    }
+  },
+  "5": {
+    "library": "models_core.json",
+    "model_name": "102_pepsi_can_12_fl_oz_vray",
+    "position": {
+      "x": -12.83,
+      "y": 1.421,
+      "z": -5.7547
+    },
+    "rotation": {
+      "x": -19.525,
+      "y": 180,
+      "z": -90
+    },
+    "scale": {
+      "x": 1.5,
+      "y": 1.5,
+      "z": 1.5
+    },
+    "physics": {
+      "kinematic": false,
+      "dynamic_friction": 0.2,
+      "static_friction": 0.2,
+      "bounciness": 0.6,
+      "mass": 0.2
+    }
+  },
+  "6": {
+    "library": "models_core.json",
+    "model_name": "h-shape_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3719,
+      "z": -6.2133
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 0.5
+    }
+  },
+  "7": {
+    "library": "models_core.json",
+    "model_name": "half_circle_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3719,
+      "z": -6.67
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1.2,
+      "z": 0.5
+    }
+  },
+  "8": {
+    "library": "models_core.json",
+    "model_name": "pentagon_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3719,
+      "z": -6.5925
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 0.5
+    }
+  },
+  "9": {
+    "library": "models_core.json",
+    "model_name": "t-shape_wood_block",
+    "position": {
+      "x": -12.95,
+      "y": 1.3719,
+      "z": -6.54
+    },
+    "rotation": {
+      "x": 0,
+      "y": 0,
+      "z": 0
+    },
+    "scale": {
+      "x": 1,
+      "y": 1,
+      "z": 0.5
+    }
+  }
+}
