+++ Python/tdw/tdw_utils.py
@@ -17,9 +17,6 @@ from tdw.librarian import ModelRecord
 from pathlib import Path
 import boto3
 from botocore.exceptions import ProfileNotFound, ClientError
-from subprocess import check_output, Popen, call
-import re
-from psutil import pid_exists
 import base64
 
 
@@ -844,99 +841,6 @@ class TDWUtils:
 
         # Source: https://github.com/Unity-Technologies/URDF-Importer/blob/c41208565419b04907496baa93ad1b675d41dc20/com.unity.robotics.urdf-importer/Runtime/Extensions/TransformExtensions.cs#L85-L92
         return np.radians(np.array([-euler_angles[2], euler_angles[0], -euler_angles[1]]))
-
-
-class AudioUtils:
-    """
-    Utility class for recording audio in TDW using [fmedia](https://stsaz.github.io/fmedia/).
-
-    Usage:
-
-    ```python
-    from tdw.tdw_utils import AudioUtils
-    from tdw.controller import Controller
-
-    c = Controller()
-
-    initialize_trial()  # Your code here.
-
-    # Begin recording audio. Automatically stop recording at 10 seconds.
-    AudioUtils.start(output_path="path/to/file.wav", until=(0, 10))
-
-    do_trial()  # Your code here.
-
-    # Stop recording.
-    AudioUtils.stop()
-    ```
-    """
-
-    # The process ID of the audio recorder.
-    RECORDER_PID: Optional[int] = None
-    # The audio capture device.
-    DEVICE: Optional[str] = None
-
-    @staticmethod
-    def get_system_audio_device() -> str:
-        """
-        :return: The audio device that can be used to capture system audio.
-        """
-
-        devices = check_output(["fmedia", "--list-dev"]).decode("utf-8").split("Capture:")[1]
-        dev_search = re.search("device #(.*): Stereo Mix", devices, flags=re.MULTILINE)
-        assert dev_search is not None, "No suitable audio capture device found:\n" + devices
-        return dev_search.group(1)
-
-    @staticmethod
-    def start(output_path: Union[str, Path], until: Optional[Tuple[int, int]] = None) -> None:
-        """
-        Start recording audio.
-
-        :param output_path: The path to the output file.
-        :param until: If not None, fmedia will record until `minutes:seconds`. The value must be a tuple of 2 integers. If None, fmedia will record until you send `AudioUtils.stop()`.
-        """
-
-        if isinstance(output_path, str):
-            p = Path(output_path).resolve()
-        else:
-            p = output_path
-
-        # Create the directory.
-        if not p.parent.exists():
-            p.parent.mkdir(parents=True)
-
-        # Set the capture device.
-        if AudioUtils.DEVICE is None:
-            AudioUtils.DEVICE = AudioUtils.get_system_audio_device()
-        fmedia_call = ["fmedia",
-                       "--record",
-                       f"--dev-capture={AudioUtils.DEVICE}",
-                       f"--out={str(p.resolve())}",
-                       "--globcmd=listen"]
-        # Automatically stop recording.
-        if until is not None:
-            fmedia_call.append(f"--until={TDWUtils.zero_padding(until[0], 2)}:{TDWUtils.zero_padding(until[1], 2)}")
-        with open(os.devnull, "w+") as f:
-            AudioUtils.RECORDER_PID = Popen(fmedia_call,
-                                            stderr=f).pid
-
-    @staticmethod
-    def stop() -> None:
-        """
-        Stop recording audio (if any fmedia process is running).
-        """
-
-        if AudioUtils.RECORDER_PID is not None:
-            with open(os.devnull, "w+") as f:
-                call(['fmedia', '--globcmd=quit'], stderr=f, stdout=f)
-            AudioUtils.RECORDER_PID = None
-
-    @staticmethod
-    def is_recording() -> bool:
-        """
-        :return: True if the fmedia recording process still exists.
-        """
-
-        return AudioUtils.RECORDER_PID is not None and pid_exists(AudioUtils.RECORDER_PID)
 
 
 class QuaternionUtils:
